feat.v1.login: implementación inicial de login móvil conectado a BD

+++ backend/demo/src/main/java/com/example/demo/usuario/repository/UserRepository.java
@@ -1,12 +1,10 @@
 package com.example.demo.usuario.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.example.demo.usuario.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> { // Cambié Long por Integer
     User findByEmailAndPassword(String email, String password);
 }
